modularized the integration-level tests to their own folder
